reset up chain instrument for rake
