+++ test/multiverse/suites/rake/Envfile
@@ -22,6 +22,8 @@
 #   gem 'rake', '0.9.6' # Newest unsupported version
 # RB
 
+instrumentation_methods :chain#, :prepend
+
 if RUBY_VERSION >= '2.2.0'
   gemfile <<-RB
     gem 'rack'
