+++ test/multiverse/suites/rake/config/newrelic.yml
@@ -6,6 +6,8 @@ development:
   rake:
     tasks:
       - "default"
+  instrumentation:
+    rake_instrumentation: <%= $instrumentation_method %>
   monitor_mode: true
   license_key: bootstrap_newrelic_admin_license_key_000
   ca_bundle_path: ../../../config/test.cert.crt
